api/pyproject.toml
[project]
name = "open-deepwiki-api"
version = "1.0.0"
description = "Backend API for DeepWiki, providing smart code analysis and AI-powered documentation generation."
license = {text = "MIT License"}

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.95.0"
psycopg = { extras = ["binary"], version = "^3.1" }
psycopg2-binary = "^2.9"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
uvicorn = { extras = ["standard"], version = ">=0.21.1" }
python-jose = "^3.3.0"
resend = "^2.0.0"
pydantic = ">=2.0.0"
google-generativeai = ">=0.3.0"
tiktoken = ">=0.5.0"
adalflow = ">=0.1.0"
numpy = ">=1.24.0"
faiss-cpu = ">=1.7.4"
langid = ">=1.1.6"
requests = ">=2.28.0"
jinja2 = ">=3.1.2"
python-dotenv = ">=1.0.0"
openai = ">=1.76.2"
ollama = ">=0.4.8"
aiohttp = ">=3.8.4"
boto3 = ">=1.34.0"
websockets = ">=11.0.3"
azure-identity = ">=1.12.0"
azure-core = ">=1.24.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
